fix/refactored large file verificationSchemaList
